requirements.txt
asgiref==3.8.1
Django==4.2.8
djangorestframework==3.14.0
et_xmlfile==2.0.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
